requirements.txt
-i https://pypi.org/simple
--extra-index-url https://download.pytorch.org/whl/cu121
aiohappyeyeballs==2.6.1; python_version >= '3.9'
aiohttp==3.12.15
aiosignal==1.4.0; python_version >= '3.9'
asttokens==3.0.0; python_version >= '3.8'
attrs==25.3.0; python_version >= '3.8'
black[jupyter]==25.1.0; python_version >= '3.9'
boto3==1.39.15; python_version >= '3.9'
botocore==1.39.15; python_version >= '3.9'
certifi==2025.7.14; python_version >= '3.7'
charset-normalizer==3.4.2; python_version >= '3.7'
click==8.2.1; python_version >= '3.10'
contourpy==1.3.3; python_version >= '3.11'
cycler==0.12.1; python_version >= '3.8'
datasets==4.0.0; python_full_version >= '3.9.0'
decorator==5.2.1; python_version >= '3.8'
dill==0.3.8; python_version >= '3.8'
executing==2.2.0; python_version >= '3.8'
filelock==3.18.0; python_version >= '3.9'
fonttools==4.59.0; python_version >= '3.9'
frozenlist==1.7.0; python_version >= '3.9'
fsspec[http]==2025.3.0; python_version >= '3.8'
pynvml>=11.0.0
hdbscan==0.8.40
hf-xet==1.1.5; platform_machine == 'x86_64' or platform_machine == 'amd64' or platform_machine == 'arm64' or platform_machine == 'aarch64'
huggingface-hub==0.34.3; python_full_version >= '3.8.0'
idna==3.10; python_version >= '3.6'
jedi==0.19.2; python_version >= '3.6'
jinja2==3.1.6; python_version >= '3.7'
jmespath==1.0.1; python_version >= '3.7'
joblib==1.5.1; python_version >= '3.9'
kiwisolver==1.4.8; python_version >= '3.10'
llvmlite==0.44.0; python_version >= '3.10'
markdown-it-py==3.0.0; python_version >= '3.8'
markupsafe==3.0.2; python_version >= '3.9'
matplotlib==3.10.3; python_version >= '3.10'
matplotlib-inline==0.1.7; python_version >= '3.8'
mdurl==0.1.2; python_version >= '3.7'
mpmath==1.3.0
multidict==6.6.3; python_version >= '3.9'
multiprocess==0.70.16; python_version >= '3.8'
mypy-extensions==1.1.0; python_version >= '3.8'
networkx==3.5; python_version >= '3.11'
numba==0.61.2; python_version >= '3.10'
numpy==2.2.6; python_version >= '3.10'
packaging==25.0; python_version >= '3.8'
pandas==2.3.1; python_version >= '3.9'
parso==0.8.4; python_version >= '3.6'
pathspec==0.12.1; python_version >= '3.8'
pexpect==4.9.0; sys_platform != 'win32' and sys_platform != 'emscripten'
pillow==11.3.0; python_version >= '3.9'
platformdirs==4.3.8; python_version >= '3.9'
prompt-toolkit==3.0.51; python_version >= '3.8'
propcache==0.3.2; python_version >= '3.9'
psutil==7.0.0; python_version >= '3.6'
ptyprocess==0.7.0
pure-eval==0.2.3
pyarrow==21.0.0; python_version >= '3.9'
pygments==2.19.2; python_version >= '3.8'
pynndescent==0.5.13
pyparsing==3.2.3; python_version >= '3.9'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2025.2
pyyaml==6.0.2; python_version >= '3.8'
regex==2024.11.6; python_version >= '3.8'
requests==2.32.4; python_version >= '3.8'
rich==14.1.0; python_full_version >= '3.8.0'
s3transfer==0.13.1; python_version >= '3.9'
safetensors==0.5.3; python_version >= '3.7'
scikit-learn==1.7.1; python_version >= '3.10'
scipy==1.16.1; python_version >= '3.11'
seaborn==0.13.2; python_version >= '3.8'
six==1.17.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
stack-data==0.6.3
sympy==1.14.0; python_version >= '3.9'
threadpoolctl==3.6.0; python_version >= '3.9'
tokenize-rt==6.2.0
tokenizers==0.21.4; python_version >= '3.9'
torch==2.7.1; python_full_version >= '3.9.0'
torchaudio==2.7.1
torchvision==0.22.1; python_version >= '3.9'
tqdm==4.67.1; python_version >= '3.7'
traitlets==5.14.3; python_version >= '3.8'
transformers==4.54.0; python_full_version >= '3.9.0'
typing-extensions==4.14.1; python_version >= '3.9'
tzdata==2025.2; python_version >= '2'
umap-learn==0.5.9.post2; python_version >= '3.9'
urllib3==2.5.0; python_version >= '3.10'
wcwidth==0.2.13
xxhash==3.5.0; python_version >= '3.7'
yarl==1.20.1; python_version >= '3.9'
